# String escape performances