@davidmarkclements 2016-10-28
Hey @delvedor can you show the benchmarking script, I'd like to see the strings and how they're constructed