@delvedor 2016-10-28
@davidmarkclements I've added the Gist link above, here it is :)
https://gist.github.com/delvedor/23111168392d83a4f994a26bdf7822b6