@davidmarkclements 2016-10-28
Thanks :)

What if the ratio of Unicode chars to ascii chars has an impact - in the short str is there a higher Unicode density? What if we change the position of the Unicode chars (say shift them all right by one by adding a char to the beginning, and popping off the last string then use flatstr to avoid concat having impact) 

I'd also like to know if a particular range of Unicode chars has a higher impact, would a higher plane take longer to parse? Are all the Unicode chars close together as far as code points go? 

Finally I'd like to see a test on lone surrogates, what effect do they have on perf and does the custom function work with lone surrogates