@delvedor 2016-10-28
@davidmarkclements I've made this script to generate the string with the unicode chars frequency. Maybe can help! :)
https://gist.github.com/delvedor/fc3e29224f1410dcd584c00811c18342