@davidmarkclements 2016-10-28
oh wait - these are *control* characters not unicode characters

everything I said about code points/code planes and lone surrogates doesn't apply